ISSUE #1735: Refactor separate combined charts to one chart
