+++ src/Domain/Activity/Stream/CombinedStream/CombinedStreamProfileCharts.php
@@ -4,6 +4,10 @@ declare(strict_types=1);
 
 namespace App\Domain\Activity\Stream\CombinedStream;
 
+use App\Infrastructure\Theme\Theme;
+use App\Infrastructure\ValueObject\Measurement\UnitSystem;
+use Symfony\Contracts\Translation\TranslatorInterface;
+
 final readonly class CombinedStreamProfileCharts
 {
     private const int GRID_HEIGHT = 120;
@@ -12,24 +16,52 @@ final readonly class CombinedStreamProfileCharts
     private const int GAP = 5;
 
     /**
-     * @param list<CombinedStreamProfileChart> $charts
+     * @param list<array{
+     *     yAxisData: array<int, int|float>,
+     *     yAxisStreamType: CombinedStreamType,
+     * }> $items
+     * @param array<int, int|float> $topXAxisData
+     * @param array<int, int|float> $bottomXAxisData
      */
     private function __construct(
-        private array $charts,
+        private array $items,
+        private array $topXAxisData,
+        private array $bottomXAxisData,
+        private int $maximumNumberOfDigitsOnYAxis,
+        private UnitSystem $unitSystem,
+        private TranslatorInterface $translator,
     ) {
     }
 
     /**
-     * @param list<CombinedStreamProfileChart> $charts
+     * @param list<array{
+     *     yAxisData: array<int, int|float>,
+     *     yAxisStreamType: CombinedStreamType,
+     * }> $items
+     * @param array<int, int|float> $topXAxisData
+     * @param array<int, int|float> $bottomXAxisData
      */
-    public static function fromCharts(array $charts): self
-    {
-        return new self($charts);
+    public static function create(
+        array $items,
+        array $topXAxisData,
+        array $bottomXAxisData,
+        int $maximumNumberOfDigitsOnYAxis,
+        UnitSystem $unitSystem,
+        TranslatorInterface $translator,
+    ): self {
+        return new self(
+            items: $items,
+            topXAxisData: $topXAxisData,
+            bottomXAxisData: $bottomXAxisData,
+            maximumNumberOfDigitsOnYAxis: $maximumNumberOfDigitsOnYAxis,
+            unitSystem: $unitSystem,
+            translator: $translator,
+        );
     }
 
     public function getTotalHeight(): int
     {
-        $count = count($this->charts);
+        $count = count($this->items);
 
         return self::TOP_MARGIN + $count * self::GRID_HEIGHT + ($count - 1) * self::GAP + self::BOTTOM_MARGIN;
     }
@@ -39,21 +71,142 @@ final readonly class CombinedStreamProfileCharts
      */
     public function build(): array
     {
-        $count = count($this->charts);
+        $count = count($this->items);
         $grids = [];
         $xAxes = [];
         $yAxes = [];
         $series = [];
         $xAxisIndices = [];
 
-        foreach ($this->charts as $index => $chart) {
-            $partialConfig = $chart->buildForCombined($index, $count);
+        $gridLeft = match ($this->maximumNumberOfDigitsOnYAxis) {
+            4 => '75px',
+            5 => '85px',
+            default => '65px',
+        };
 
-            $grids[] = $partialConfig['grid'];
-            $xAxes[] = $partialConfig['xAxis'];
-            $yAxes[] = $partialConfig['yAxis'];
-            $series[] = $partialConfig['series'];
+        foreach ($this->items as $index => $item) {
+            $yAxisData = $item['yAxisData'];
+            $yAxisStreamType = $item['yAxisStreamType'];
+            $isFirst = 0 === $index;
+            $isLast = $index === $count - 1;
+
+            if ([] === $yAxisData) {
+                throw new \RuntimeException('yAxisData data cannot be empty');
+            }
+
+            $yAxisSuffix = $yAxisStreamType->getSuffix($this->unitSystem);
+
+            [$min, $max] = [min($yAxisData), max($yAxisData)];
+            $margin = ($max - $min) * 0.1;
+            $maxYAxis = (int) ceil($max + $margin);
+            $minYAxis = max($min, 0);
+
+            if (CombinedStreamType::ALTITUDE === $yAxisStreamType) {
+                $minYAxis = (int) floor($min - $margin);
+            }
+
+            $top = self::TOP_MARGIN + $index * (self::GRID_HEIGHT + self::GAP);
+
+            $grids[] = [
+                'left' => $gridLeft,
+                'right' => '20px',
+                'top' => $top.'px',
+                'height' => self::GRID_HEIGHT.'px',
+                'containLabel' => false,
+            ];
+
+            $xAxisPosition = match (true) {
+                $isFirst => Theme::POSITION_TOP,
+                $isLast => Theme::POSITION_BOTTOM,
+                default => null,
+            };
+            $xAxisData = match (true) {
+                $isLast && [] !== $this->bottomXAxisData => $this->bottomXAxisData,
+                default => $this->topXAxisData,
+            };
+            $xAxisLabelSuffix = match (true) {
+                $isLast && [] !== $this->bottomXAxisData => $this->unitSystem->distanceSymbol(),
+                default => null,
+            };
+
+            $xAxes[] = [
+                'gridIndex' => $index,
+                'position' => $xAxisPosition,
+                'type' => 'category',
+                'boundaryGap' => false,
+                'axisLabel' => [
+                    'show' => !is_null($xAxisPosition),
+                    'formatter' => '{value} '.$xAxisLabelSuffix,
+                ],
+                'data' => $xAxisData,
+                'min' => 0,
+                'axisTick' => [
+                    'show' => false,
+                ],
+            ];
             $xAxisIndices[] = $index;
+
+            $yAxes[] = [
+                'gridIndex' => $index,
+                'type' => 'value',
+                'name' => $yAxisStreamType->trans($this->translator),
+                'nameRotate' => 90,
+                'nameLocation' => 'middle',
+                'nameGap' => 10,
+                'min' => $minYAxis,
+                'max' => $maxYAxis,
+                'splitLine' => [
+                    'show' => false,
+                ],
+                'axisLabel' => [
+                    'show' => true,
+                    'customValues' => [$minYAxis, $maxYAxis],
+                    'color' => '#aaa',
+                    'verticalAlignMaxLabel' => 'top',
+                    'verticalAlignMinLabel' => 'bottom',
+                ],
+            ];
+
+            $series[] = [
+                'name' => CombinedStreamType::PACE === $yAxisStreamType ? '__pace' : $yAxisSuffix,
+                'xAxisIndex' => $index,
+                'yAxisIndex' => $index,
+                'markArea' => [
+                    'data' => [
+                        [
+                            [
+                                'xAxis' => 'min',
+                                'itemStyle' => [
+                                    'color' => '#3E444D',
+                                ],
+                                'emphasis' => [
+                                    'disabled' => true,
+                                ],
+                            ],
+                            [
+                                'xAxis' => 'max',
+                            ],
+                        ],
+                    ],
+                ],
+                'data' => $yAxisData,
+                'type' => 'line',
+                'showSymbol' => false,
+                'progressive' => 5000,
+                'progressiveThreshold' => 10000,
+                'color' => $yAxisStreamType->getSeriesColor(),
+                'smooth' => true,
+                'lineStyle' => [
+                    'width' => 0,
+                ],
+                'emphasis' => [
+                    'disabled' => true,
+                ],
+                'areaStyle' => [
+                    'opacity' => 1,
+                    'origin' => 'start',
+                ],
+            ];
         }
 
         return [
